Make SensorSDK code even more generic.
SensorSDK now provides a base code which handles:
* Button interrupts
* ALARM interrupt
* INIT interrupt
* LCD displaying
* deep sleep functions (comming soon)
* Configuration Menu (including inquirying for pairing)
* User Extension Points (including insertion point for sensor reading)

On top of SensorSDK we implement 2 main different codes monitor and interactive.
We understand for a monitor node a monitor that periodically meassures a sensor
and somehow reports this values. An interactive node is a more human friendly
node in the sense it will allow the user to take readings and then take
different actions depending on the values that are readed. The user is provided
with a menu which can be configured on runtime.

This new approach requires changes to all the other overlays, next commit will
include monitor overlay, and the other one will include sensor overlays